Refactor Utils.java
