+++ app/src/main/java/nil/nadph/qnotified/hook/InspectMessage.java
@@ -39,7 +39,7 @@ import nil.nadph.qnotified.util.Utils;
 
 import static de.robv.android.xposed.XposedHelpers.findAndHookMethod;
 import static nil.nadph.qnotified.util.Initiator.load;
-import static nil.nadph.qnotified.util.ReflexUtil.iget_object_or_null;
+import static nil.nadph.qnotified.util.ReflexUtil.*;
 import static nil.nadph.qnotified.util.Utils.*;
 
 public class InspectMessage extends CommonDelayableHook implements View.OnLongClickListener {
