+++ app/src/main/java/nil/nadph/qnotified/hook/MultiForwardAvatarHook.java
@@ -37,8 +37,7 @@ import nil.nadph.qnotified.util.*;
 
 import static de.robv.android.xposed.XposedHelpers.findAndHookMethod;
 import static nil.nadph.qnotified.util.Initiator.load;
-import static nil.nadph.qnotified.util.ReflexUtil.iget_object_or_null;
-import static nil.nadph.qnotified.util.ReflexUtil.invoke_static_any;
+import static nil.nadph.qnotified.util.ReflexUtil.*;
 import static nil.nadph.qnotified.util.Utils.*;
 
 public class MultiForwardAvatarHook extends CommonDelayableHook {
